Add Linux script to limit outgoing bandwidth used by Bitcoin
This commit adds a Linux script that uses tc and iptables to
limit outgoing bandwidth to Bitcoin nodes. All tcp connections with
a source or destination port of 8333 are limited to the chosen rate.
It does not alter the incoming bandwidth. Additionally, outgoing
bandwidth to a host on a local LAN (defined by the LOCALNET variable)
is not limited.


Former-commit-id: 2002bceac5b8fad593c9cf83cc4e7127029f2174