minimum collateral amount
In the case where collateral is made as COLLATERAL*5, then
used 4 times, it will result in an input exactly COLLATERAL sized.
When used as a collateral transaction this will result in 1 input of
COLLATERAL and zero outputs (due to collateral being paid solely in fees).
zero output transactions aren't supported, so it'll result in DS failure.


Former-commit-id: b5f6f0e7a735eb44e4062493ed6296b75c07ef79