[Qt] Refactor layout for rpcconsole UI
Improve readability of the `rpcconsole.ui` form file by reordering
layout elements for the Information and Wallet Repair tabs. Also
convert the Wallet Repair tab to a QGridLayout format similar to
how the Information tab is.

Benefit of using the QGridLayout is that we no longer pass the
repair commands in the description string to Transifex, which we
shouldn't have been doing anyways.


Former-commit-id: 9d712951182ca2dce428a7c665331f49286ef8ec [formerly 387cce9b93a762139708fc0ed680947dc2f9d75e]
Former-commit-id: 03d51e024caef6f45c59b46a8f1acaaa26c715ba
Former-commit-id: f7f6a3e98a53d068585dcb00243a30d90b34c272