[Qt] Adjust size of splash screen image.
Add extra height to avoid having the initialization status text
overlap with the copyright text.

Since the splash window size is based on the static width/height of
the pre-rendered image files, just add an extra 10px to the image
height.

Also made the background for the testnet splash image uniform in
color.


Former-commit-id: 1d65e871632b3d9c29998cff12a6d6beff51682c [formerly fea5a1c052d0856549339730a0e127a45bf1bfe3]
Former-commit-id: e39dd72cf3506db80dc9d5e83137d200da040687
Former-commit-id: ba3ff9956ddc558758e234ebf4e25340590a357f