[Qt] Fix segfault when running with `-help`
When running pivx-qt with the `-help` flag, the intended operation
is to output the help contents then exit cleanly. However, this is
currently resulting in a segfault due to the return status of a
method in intro.

This Fixes the segfaulting behavior and now cleanly exits after
displaying the help contents.


Former-commit-id: 41693a17a34f5f2c304edfa56dbf037671d5e66f [formerly 37d2f1b1308a0c07fc2ef21c9c9148fc4dc1deb9]
Former-commit-id: c90acfafee02da3933b05e6ba8cd0037f0c53ff3
Former-commit-id: 5ee498891601692c681b22db00b9e81a9ac80d71