zPiv mints with < 3 mints added to accumulator after are immature.
To generate a witness, and therefore spend, a zPiv mint must have at least 3 mints added to the accumulator after it is accumulated itself. If that condition is not met, then the spender would receive the error message "Less than 3 mints added, unable to create spend".

Now, ListMintedCoins() will check to see if the mint has at least 3 mints after it before considering it mature and ready for spending.


Former-commit-id: 3a16c3a077bb162e4b2b4614812cb96187b9277c [formerly 22ad2ec81d826fcc4d419377b0cb6470e9123313]
Former-commit-id: 0a321509324d4e37240970697a38567f63a70faf
Former-commit-id: e03239195aa34a68598d96517fb1fef1a065e142