Application shutdown request interrupts missing accumulators process
When user would try to quit PIVX in the middle of the "Calculating missing accumulators" process, the application would hang for up to 30 min (reportedly). Additionally, the request to close the application appeared to be completely ignored by AppInit2. These changes make the GetCheckpoint process check for shutdown requests between block verifications, which allows the application to quit faster. Additionally, the bool vars that controlled whether a shutdown request had been sent were changed to volitile so that compilers wouldn't optimize out the if statements responsible for catching a shutdown request.


Former-commit-id: fde0e4728737296b7f4171710f4f651d4d4a4af7 [formerly 1e17a395a956a4f0ef709b288c2492ab5455acde]
Former-commit-id: f83db1c7614bfe48bef8dcb26bce10ba6fe32260
Former-commit-id: f7867e935362386976af684751ed01fb67216594