Do not stringify anything beyond OP_ZEROCOINSPEND for CScript::ToString()
There are no further op codes in zerocoinspends. When parsing through the data, the ToString() method often see's the zPiv spend data as an OP code. Since RPC getrawtransaction prints the hex of the CScript anyways, it is redundant to print all of the spend data in this to string and then again in the hex dump.


Former-commit-id: b26eea794d73025d903007ed398c6432196e0bd9 [formerly 1297a80e2002d69464be0049d4c7701a80435bee]
Former-commit-id: c50a5bbe07387680927411685cd35498b52b74b3
Former-commit-id: a6a7127ba544de887bb419ca51db7b35c009a4c2